walmart/app/requirements.txt
termcolor
#Pyqt5
#Pmw
#pyserial >= 3.4
Django == 2.2.12
psycopg2
djangorestframework
django-crispy-forms
django-braces
channels == 1.1.8
requests
'''
ioredis            1.1.0              
asgi-redis          1.4.3              
asgiref             1.1.2              
async-timeout       2.0.1              
attrs               17.3.0             
autobahn            17.10.1            
Automat             0.6.0              
certifi             2019.11.28         
channels            1.1.8              
chardet             3.0.4              
constantly          15.1.0             
cymysql             0.9.2              
daphne              1.3.0              
dj-database-url     0.5.0              
Django              2.2.12             
django-braces       1.14.0             
django-countries    6.0                
django-crispy-forms 1.7.2              
django-cymysql      0.5.1              
djangorestframework 3.11.0             
gitdb2              2.0.3              
GitPython           2.1.9              
hiredis             0.2.0              
hyperlink           17.3.1             
idna                2.6                
incremental         17.5.0             
msgpack             0.5.6              
msgpack-python      0.4.8              
pathlib             1.0.1              
Pillow              5.3.0              
pip                 19.3.1             
Pmw                 2.0.1              
psycopg2            2.8.5              
PyQt5               5.14.2             
PyQt5-sip           12.7.2             
pyserial            3.4                
pytz                2018.4             
redis               2.10.6             
requests            2.23.0             
setuptools          44.0.0.post20200106
six                 1.11.0             
smmap2              2.0.3              
sqlparse            0.3.1              
termcolor           1.1.0              
Twisted             17.9.0             
txaio               2.8.2              
urllib3             1.25.8             
wheel               0.33.6             
zope.interface      4.4.3 
'''